// src/types/goods.ts
export type GoodsType = {
  id?: string
  describe?: string
  gain?: string
  name?: string
  resourceName?: string
  resourcePath?: string
  showBag?: string
  isExpire?: boolean
  story?: string
  extraData?: string
  count?: number
  templateId?: number
  expiredTime?: number
  picture?: string
}

export type GoodsDiaDetail = {
  img?: string
  title?: string
  desc?: string
  detail?: string
  code?: string
  type?: number
  gain?: string
}